Update registration.php
Hello Alexis and Brandon. I just made a little edit. Since it is supposed to be for a user (guest) creating an account registration for their first time.

Therefore, Instead of just labeling it as just "Password" I changed it to "Create a Password." For the Username as well, "Create a Username."  I also added the following; "Confirm Password" "Show Password function"